Added Euro Low Pass Filter to the plugin (same on that Epic is using with the VR Editor for their laser beam smoothing). It has ResetEuroFilter and RunEuroFilterSmoothing nodes in the VR Expansion Library to control it.
Also added an auto arc calculation node to the VR Expansion Function Library so that the
arc calculation and setting is faster and clean and is the same as the VR Editor.

I normally wouldn't directly port things over, however I know that people would want the
same behavior if possible so that they have consistancy.

The template contains examples of using them now.

Also removed a silly check left over from copying code in the VRBaseCharacter.
It likely had no effect but it should be removed.