Overrode default engine rollback / client correction code to now pass and verify a rotation instead of directly accepting the clients rotation every movement.
This allows me to rollback rotations and correct for packet loss with rotation based movements.
Something that is not very important in FPS games because rotation is at the zero point but
which is fairly important in offset HMD VR games as any minor rotation offset can cause pretty hefty
locational offsets.

Added bUseClientControlRotation to VRCharacterMovementComponent, this allows the movement
component to use the legacy behavior of not rolling back rotations (FPS test pawns).

These changes are for the VRCharacter only, the SimpleVRcharacter does not need these
modifications to function. They also need some testing, vehicles specifically could use some testing.



Added MoveAction_SetRotation (For just changing the rotation) and changed the snap turn and teleport
move actions to pass in the end yaw and simulate the movement instead of taking just a location offset.

Added make and conversion nodes for FVector_Netquantize and its variants...since the engine is missing this.