Raised boundry for rotation rollback a little.
Still need to remove magic number and expose a console variable for this soon

Send compressed axis for rotators is very nearly 2 decimal precision ( rounding errors
ends up with it being slightly less). Decided to bring threshold up to 0.02f to account for the
rounding errors when converting from uint16 to float again on the other side.