Converted the gesture component to use a timer instead of a manually counting tick.
Also running the debug draw lifetime off of the delta time for recording so that
there is no flicker.

This doesn't save much perf over the manual method unless you have a ton of
gesture trackers, however there is no real need to use tick here and a timer
makes more sense.

The only downside is if a frame takes longer than the update timer for the set
HTZ, the timer can trigger twice then, which is kind of a problem since we don't
want to sample the same spot multiple times, however I already take care of that
and don't include duplicate entries right after each other.