Moving NavMoveCompleted events into BaseMovementComponent so that the SimpleVRCharacter can use them.
Also re-defined GetActorFeetLocation as GetActorFeetLocationVR so that we have a special one
just for our use instead of overwriting the default one.

The one downside is that the VR version cannot be inlined since it access's a parent class.
However this function is only called in the characters navigation so it shouldn't be a big
deal, users shouldn't be doing flocking AI with a VRBaseCharacter type anyway.