added OnDestroyed back to components, but I am filtering out editor views now so that it saves correctly.
more grip script replication clean up

A bit of a hack here forcing the OnPreReplicationDestroy onto our components.
However it should be a valid correction and I needed a good event for this.

There is no default implementation anywhere for this function on components
so it should be safe to use.

Also fixed up grip scripts a bit so that they return the correct valid owning outer.