IgnoreActorWhenMoving is causing overlap spam, it really only effects the interactive with sweep and hybrid grip types so its not the largest deal ever since good collision channel management can handle it.
Its supposed to just drop overlap events like this, but obviously the execution is
bugged.

this commit removes using it to stop the spam and to check if there are any ill effects
with the user base.

PLEASE let me know if this screws over an interaction that you had going for you.

Its also possible to manually call it OnGrip on the clients side as well when needed.
Though if doing that, its likely that all sub components need to be set the same.