Moved EVRInteractibleAxis to the VRInteractibleFunctionLibrary header, it belongs there instead.
Added new grip type "EventsOnly", will skip ALL extra logic of grips and only store the grip
data and fire the events for it. Whether you do anything else with it or not is up to you.
Like a CustomGrip except even more barebones as it does not fire TickGrip or obey any of the
dropping simulation logic, it is also auto excluded from late updates.

Stopped calling "StopSimulating" on Custom Gripped objects automatically.
It is legal to start and keep a custom grip simulating, that call should have
never been there. Anyone who needs their custom grips to auto stop simulating on grip
will need to call it in the OnGrip event themselves.