Implemented the seated mode "fix" with an optional boolean to turn it off if it causes problems.
bUseExperimentalUnseatModeFix, defaulted to True, set to false if you notice issues
caused by it. If it proves stable then the boolean will be removed in the near future.

Am now overriding the ForcePositionUpdate so that it also ignores the SeatedMode custom
movement mode as well as the std behavior of ignoring the MovementMode_None mode.