Grippables and the VRGripInterface will start storing multiple controllers for grips generally most objects won't be using more than one at a time, this is pretty much specifically just for physics grips at this point
but eventually it may be used to tag "primary" hands and "Secondary" hands and manage grips
a little better, for example, adding another grip at the front of a gun and pausing it, when
the primary grip is released you just unpause the secondary one to have it take over movement.

i'll have to really go in and make sure that things like turning off gravity during a grip
and other such things don't have issues with multiple concurrent grips eventually

the main difference in peoples implementations would most just be changing to using the
new format for querying IsHeld.

I was always "posslble" to multiple grip objects, but the IsHeld node would only return
the last one to grip and it was harder to manage those special tags
I won't be implementing anything with this for awhile yet, I just wanted to get the back bone
in for future stuff and get it out of the way