Added a FlightStick_XY mode to the VRLever component. Will follow the orientation of the hand in YAW as well as the angle in Pitch/Roll.
Added the rotator property: AllCurrentLeverAngles

CurrentLeverAngle stands for the primary angle of the lever

CurrentLeverForwardVector is the directional vector that the lever is pointing in.

AllCurrentLeverAngles is a rotator with all values set, this is most useful in XY
and flightstick mode as both Roll and Pitch will have seperate angular values in it.
Also with the flightstick mode Yaw will also be filled in.

Cleaned up a lot of the lever calcs.

Switched to a quaternion lerp for XY modes post grip (more direct).

Removed YZ lever mode, it was never used and was a tacked on pull request that
was replaced with the VRMountComponent from the same author.