Added slot name retention to the grip functions and data rep by optionally allowing the end user to pass in a name for the slot being used.
On the nodes that had a SnapToSocket name already it uses this variable.

This changes the behavior with the auto snap to socket where it will only snap to
that socket if the passed in transform is empty. Otherwise it will just store
the socket name.

This also involved added that param to the GripObjectByInterface node.